requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   google-genai
    #   httpx
babel==2.18.0
    # via courlan
beautifulsoup4==4.14.3
    # via -r requirements.in
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
    #   requests
    #   trafilatura
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.6
    # via
    #   htmldate
    #   requests
    #   trafilatura
courlan==1.3.2
    # via trafilatura
cryptography==46.0.6
    # via google-auth
dateparser==1.2.2
    # via htmldate
exceptiongroup==1.3.1
    # via anyio
google-auth==2.49.1
    # via google-genai
google-genai==1.47.0
    # via -r requirements.in
h11==0.16.0
    # via httpcore
htmldate==1.9.4
    # via trafilatura
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via google-genai
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
importlib-metadata==8.7.1
    # via markdown
jinja2==3.1.6
    # via -r requirements.in
justext==3.0.2
    # via trafilatura
lxml[html-clean]==6.0.2
    # via
    #   htmldate
    #   justext
    #   lxml-html-clean
    #   trafilatura
lxml-html-clean==0.4.4
    # via lxml
markdown==3.9
    # via -r requirements.in
markupsafe==3.0.3
    # via
    #   -r requirements.in
    #   jinja2
minify-html==0.18.1
    # via -r requirements.in
pillow==11.3.0
    # via -r requirements.in
pyasn1==0.6.3
    # via pyasn1-modules
pyasn1-modules==0.4.2
    # via google-auth
pycparser==2.23
    # via cffi
pydantic==2.12.5
    # via google-genai
pydantic-core==2.41.5
    # via pydantic
python-dateutil==2.9.0.post0
    # via
    #   dateparser
    #   htmldate
python-frontmatter==1.1.0
    # via -r requirements.in
pytz==2026.1.post1
    # via dateparser
pyyaml==6.0.3
    # via
    #   -r requirements.in
    #   python-frontmatter
regex==2026.1.15
    # via dateparser
requests==2.32.5
    # via
    #   -r requirements.in
    #   google-genai
six==1.17.0
    # via python-dateutil
soupsieve==2.8.3
    # via beautifulsoup4
tenacity==9.1.2
    # via google-genai
tld==0.13.2
    # via courlan
trafilatura==2.0.0
    # via -r requirements.in
typing-extensions==4.15.0
    # via
    #   anyio
    #   beautifulsoup4
    #   cryptography
    #   exceptiongroup
    #   google-genai
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzlocal==5.3.1
    # via dateparser
urllib3==2.6.3
    # via
    #   courlan
    #   htmldate
    #   requests
    #   trafilatura
websockets==15.0.1
    # via google-genai
zipp==3.23.0
    # via importlib-metadata
